Refactor the path handling code into a Path object and use that everywhere.

+++ s3commander.js
@@ -13,7 +13,79 @@ b64pad = "=";
     "use strict";
 
     /************************************************************************
-     * General / Utility                                                    *
+     * Utility                                                              *
+     ************************************************************************/
+
+    function Path(sPath, bFolder) {
+        sPath = typeof sPath !== 'undefined' ? sPath : "";
+        bFolder = typeof bFolder !== 'undefined' ? bFolder : false;
+
+        this.parts = sPath.split("/");
+        this.folder = bFolder;
+        this.normalize();
+    }
+
+    Path.prototype.normalize = function() {
+        this.parts = this.parts.filter(function(part){
+          return part.length > 0;
+        });
+    };
+
+    Path.prototype.toString = function() {
+        var uri = this.parts.join("/");
+        if (this.folder && this.parts.length > 0) {
+            uri += "/";
+        }
+
+        return uri;
+    };
+
+    Path.prototype.clone = function() {
+        var other = new Path();
+        other.parts = new Array(this.parts);
+        other.folder = this.folder;
+
+        return other;
+    };
+
+    Path.prototype.empty = function() {
+        return this.parts.length == 0;
+    };
+
+    Path.prototype.push = function(sPath) {
+        var newparts = sPath.split("/");
+        Array.prototype.push.apply(this.parts, newparts);
+
+        this.folder = (newparts.length > 0 && sPath.substr(-1) == "/");
+        this.normalize();
+
+        return this;
+    };
+
+    Path.prototype.pop = function() {
+        this.parts.pop();
+
+        return this;
+    };
+
+    Path.prototype.extend = function(pOther) {
+        this.parts = this.parts.concat(pOther.parts);
+        this.folder = pOther.folder;
+        this.normalize();
+
+        return this;
+    };
+
+    Path.prototype.concat = function(pOther) {
+        var result = new Path();
+        result.parts = this.parts.concat(pOther.parts);
+        result.folder = pOther.folder;
+
+        return result;
+    };
+
+    /************************************************************************
+     * Internal State                                                       *
      ************************************************************************/
 
     /**
@@ -38,59 +110,6 @@ b64pad = "=";
      */
     var uiUploadControl = null;
 
-    /**
-     * Normalize a URI by removing empty components ('//') and leading slashes.
-     * If bTrim is true then it will also remove trailing slashes otherwise it
-     * keeps a single trailing slash if the original URI had one.
-     */
-    function normURI(sURI, bTrim) {
-        // default parameter values
-        bTrim = typeof bTrim !== 'undefined' ? bTrim : false;
-
-        // corner case: empty uri
-        if (sURI.length == 0) {
-            return "";
-        }
-
-        // split the uri and filter out empty components
-        var folder = (sURI.substr(-1) == "/");
-        var parts = sURI.split("/").filter(function(part){
-            return part.length > 0;
-        });
-
-        // create the uri from it's components
-        var uri = parts.join("/");
-        if (folder && !bTrim) {
-            return uri + "/";
-        }
-
-        return uri;
-    }
-
-    /**
-     * Create a normalized URI from components.
-     */
-    function joinURI(aParts, bTrim) {
-        // default parameter values
-        bTrim = typeof bTrim !== 'undefined' ? bTrim : false;
-
-        // join the parts and normalize the result
-        return normURI(aParts.join("/"), bTrim);
-    }
-
-    /**
-     * Pop the last component from the given URI and return the remaining part.
-     */
-    function popURI(sURI) {
-        // corner case: empty uri
-        if (sURI.length == 0) {
-            return sURI;
-        }
-
-        // remove everything starting from the last '/'
-        return sURI.substr(0, sURI.lastIndexOf("/"));
-    }
-
     /************************************************************************
      * Amazon AWS                                                           *
      ************************************************************************/
@@ -107,14 +126,11 @@ b64pad = "=";
      *
      * http://docs.aws.amazon.com/AmazonS3/latest/dev/RESTAuthentication.html
      */
-    function signRequest(opts, sMethod, sResource, oParams) {
+    function signRequest(opts, sMethod, pResource, oParams) {
         // default parameter values
         sMethod = typeof sMethod !== 'undefined' ? sMethod : "GET";
-        sResource = typeof sResource !== 'undefined' ? sResource : "";
+        pResource = typeof pResource !== 'undefined' ? pResource : new Path();
         oParams = typeof oParams !== 'undefined' ? oParams : new Object();
-
-        // normalize the resource
-        sResource = normURI(sResource);
 
         // this is used as the request and signature expiration timestamp
         // for convenince. the request timestamp must be within 15
@@ -127,8 +143,8 @@ b64pad = "=";
         secure += timestamp + "\n";
         secure += "/" + opts.sBucket + "/";
 
-        if (sResource.length > 0) {
-            secure += normURI(opts.sPrefix + "/" + sResource);
+        if (!pResource.empty()) {
+            secure += opts.pPrefix.concat(pResource).toString();
         }
 
         var params = $.param(oParams);
@@ -155,8 +171,8 @@ b64pad = "=";
     /**
      * Retrieve a url for the given resource.
      */
-    function getResourceUrl(opts, sResource) {
-        return getAPIUrl(opts) + "/" + joinURI([opts.sPrefix, sResource]);
+    function getResourceUrl(opts, pResource) {
+        return getAPIUrl(opts) + "/" + opts.pPrefix.concat(pResource).toString();
     }
 
     /**
@@ -169,7 +185,7 @@ b64pad = "=";
             "conditions": [
                 {"acl": "private"},
                 {"bucket": opts.sBucket},
-                ["starts-with", "$key", opts.sPrefix],
+                ["starts-with", "$key", opts.pPrefix.toString()],
                 ["starts-with", "$Content-Type", ""],
             ],
         };
@@ -187,7 +203,7 @@ b64pad = "=";
     }
 
     /************************************************************************
-     * Plugin Actions                                                       *
+     * S3 Backend                                                           *
      ************************************************************************/
 
     /**
@@ -195,19 +211,20 @@ b64pad = "=";
      *
      * http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketGET.html
      */
-    function listContents(sPath) {
+    function listContents(pLocation) {
         var opts = container.data("opts");
 
         // default parameter values
-        sPath = typeof sPath !== 'undefined' ? sPath : "";
+        pLocation = typeof pLocation !== 'undefined' ? pLocation : new Path("", true);
 
         // determine the full path and sign the request
-        var fullpath = joinURI([opts.sPrefix, sPath], true);
-        if (fullpath.length > 0) {
-            fullpath += "/";
+        if (!pLocation.folder) {
+            console.log("listContents(): not a folder: " + pLocation.toString());
+            return;
         }
 
-        var signdata = signRequest(opts, "GET", "");
+        var fullpath = opts.pPrefix.concat(pLocation).toString();
+        var signdata = signRequest(opts, "GET", new Path("", true));
 
         // request bucket contents with the given prefix and group results
         // into common prefixes using a delimiter
@@ -225,7 +242,8 @@ b64pad = "=";
                 var folders = $(data).find("ListBucketResult > CommonPrefixes > Prefix");
 
                 function extract(e){
-                    return normURI(e.innerHTML.substr(fullpath.length), true);
+                    var relpath = e.innerHTML.substr(fullpath.length);
+                    return new Path(relpath).toString();
                 }
 
                 function keep(e){
@@ -233,7 +251,7 @@ b64pad = "=";
                 }
 
                 container.data("contents", {
-                    "path": normURI(sPath, true),
+                    "path": pLocation,
                     "files": $.map(files, extract).filter(keep),
                     "folders": $.map(folders, extract).filter(keep),
                 });
@@ -250,13 +268,18 @@ b64pad = "=";
      * Create a folder with the given path. Folders are S3 objects where
      * the key ends in a trailing slash.
      */
-    function createFolder(sPath) {
+    function createFolder(pResource) {
         var opts = container.data("opts");
-        sPath = normURI(sPath, true) + "/";
-        var signdata = signRequest(opts, "PUT", sPath);
+        if (!pResource.folder) {
+            console.log("createFolder(): not a folder: " + pResource.toString());
+            return;
+        }
+
+        var signdata = signRequest(opts, "PUT", pResource);
+        var url = getResourceUrl(opts, pResource) + "?" + $.param(signdata);
 
         return $.ajax({
-            url: getResourceUrl(opts, sPath) + "?" + $.param(signdata),
+            url: url,
             type: "PUT",
             data: "",
             error: function(data){
@@ -270,13 +293,18 @@ b64pad = "=";
      * Delete the folder at the given path. Folders are S3 objects where
      * the key ends in a trailing slash.
      */
-     function deleteFolder(sPath) {
+     function deleteFolder(pResource) {
          var opts = container.data("opts");
-         sPath = normURI(sPath, true) + "/";
-         var signdata = signRequest(opts, "DELETE", sPath);
+         if (!pResource.folder) {
+           console.log("deleteFolder(): not a folder: " + pResource.toString());
+           return;
+         }
+
+         var signdata = signRequest(opts, "DELETE", pResource);
+         var url = getResourceUrl(opts, pResource) + "?" + $.param(signdata);
 
          return $.ajax({
-             url: getResourceUrl(opts, sPath) + "?" + $.param(signdata),
+             url: url,
              type: "DELETE",
              error: function(data){
                  console.log("Error: " + data.responseText);
@@ -291,14 +319,19 @@ b64pad = "=";
      *
      * http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html
      */
-    function downloadFile(sPath) {
+    function downloadFile(pResource) {
         var opts = container.data("opts");
-        var signdata = signRequest(opts, "GET", sPath, {
+        if (pResource.folder) {
+          console.log("downloadFile(): not a file: " + pResource.toString());
+          return;
+        }
+
+        var signdata = signRequest(opts, "GET", pResource, {
             'response-cache-control': 'No-cache',
             'response-content-disposition': 'attachment'
         });
 
-        var url = getResourceUrl(opts, sPath) + "?" + $.param(signdata);
+        var url = getResourceUrl(opts, pResource) + "?" + $.param(signdata);
         window.open(url, "_blank");
     }
 
@@ -307,11 +340,15 @@ b64pad = "=";
      *
      * http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectDELETE.html
      */
-    function deleteFile(sPath) {
+    function deleteFile(pResource) {
         var opts = container.data("opts");
+        if (pResource.folder) {
+          console.log("deleteFile(): not a file: " + pResource.toString());
+          return;
+        }
 
-        var signdata = signRequest(opts, "DELETE", sPath);
-        var url = getResourceUrl(opts, sPath) + "?" + $.param(signdata);
+        var signdata = signRequest(opts, "DELETE", pResource);
+        var url = getResourceUrl(opts, pResource) + "?" + $.param(signdata);
 
         return $.ajax({
             url: url,
@@ -363,8 +400,7 @@ b64pad = "=";
             .html("Up")
             .click(function(){
                 var contents = container.data("contents");
-                var path = popURI(contents.path);
-                listContents(path).then(updateDisplay);
+                listContents(contents.path.pop()).then(updateDisplay);
             })
             .appendTo(uiBreadcrumbs);
     }
@@ -383,7 +419,7 @@ b64pad = "=";
 
       // create crumbs
       var uiButton = uiBreadcrumbs.find("button[role='s3c-crumb-refresh']");
-      $.each(contents.path.split("/"), function(i, crumb){
+      $.each(contents.path.parts, function(i, crumb){
           $("<span />")
               .attr("role", "s3c-crumb-sep")
               .html("/")
@@ -446,7 +482,7 @@ b64pad = "=";
 
                 // create the folder
                 var contents = container.data("contents");
-                createFolder(joinURI([contents.path, name]))
+                createFolder(contents.path.clone().push(name + "/"))
                     .then(function(){ return listContents(contents.path); })
                     .then(updateDisplay);
 
@@ -556,7 +592,7 @@ b64pad = "=";
       var amazondata = $.extend(getPolicyData(opts), {
           "AWSAccessKeyId": opts.sAccessKey,
           "Content-Type": "application/octet-stream",
-          "key": joinURI([opts.sPrefix, contents.path, "${filename}"], true),
+          "key": opts.pPrefix.concat(contents.path).push("${filename}").toString(),
       });
 
       uiUploadControl.find("input[role='s3c-upload-setting']").remove();
@@ -602,7 +638,7 @@ b64pad = "=";
         // create folder entries
         container.find("div[role='s3c-folder']").remove();
         $.each(contents.folders, function(i, folder){
-            var path = contents.path + "/" + folder;
+            var path = contents.path.clone().push(folder + "/");
             var entry = $("<div />")
                 .attr("role", "s3c-folder")
                 .addClass(opts.entryClasses.join(" "));
@@ -634,7 +670,7 @@ b64pad = "=";
         // create file entries
         container.find("div[role='s3c-file']").remove();
         $.each(contents.files, function(i, file){
-            var path = contents.path + "/" + file;
+            var path = contents.path.clone().push(file);
             var entry = $("<div />")
                 .attr("role", "s3c-file")
                 .addClass(opts.entryClasses.join(" "));
@@ -665,7 +701,7 @@ b64pad = "=";
     }
 
     /************************************************************************
-     * jQuery                                                               *
+     * jQuery Plugin                                                        *
      ************************************************************************/
 
     // create an s3commander window
@@ -679,13 +715,13 @@ b64pad = "=";
             options));
 
         var opts = container.data("opts");
-        opts.sPrefix = normURI(opts.sPrefix, true);
+        opts.pPrefix = new Path(opts.sPrefix, true);
 
         // create the display
         createDisplay();
 
         // get the contents of the top-level folder
-        container.data("contents", {"path": ""});
+        container.data("contents", {});
         listContents().then(updateDisplay);
 
         // return the container
@@ -705,4 +741,12 @@ b64pad = "=";
         "formClasses": ["s3form", "form-inline"],
         "buttonClasses": ["btn", "btn-xs", "btn-primary", "pull-right"],
     };
+
+    /************************************************************************
+     * Debug                                                                *
+     ************************************************************************/
+
+    // export objects
+    window.Path = Path;
+
 }(jQuery));
